# apps/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "get": {
        "description": "Renvoie tous les clients de l'organisation du marchand authentifié. Filtres possibles : recherche, type de client et statut d'activité.",
        "operationId": "CustomersController_findAll",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numéro de page",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtrer par activité (active = au moins une transaction, inactive = aucune transaction)",
            "schema": {
              "enum": ["active", "inactive", "all"],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filtrer par type de client",
            "schema": {
              "enum": ["business", "individual", "all"],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche par nom ou e-mail",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée de clients",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "customers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CustomerResponseDto"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "pageSize": {
                          "type": "number",
                          "example": 50
                        },
                        "totalCount": {
                          "type": "number",
                          "example": 100
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 2
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lister les clients",
        "tags": ["Clients"]
      },
      "post": {
        "description": "Crée un client dans votre organisation. Il est automatiquement rattaché à votre organisation.",
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Jane Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "whatsapp_number": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "postal_code": {
                    "type": "string"
                  },
                  "is_business": {
                    "type": "boolean"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              },
              "examples": {
                "minimal": {
                  "summary": "Client minimal",
                  "value": {
                    "name": "Jane Doe",
                    "email": "jane@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données d'entrée invalides"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Créer un client",
        "tags": ["Clients"]
      }
    },
    "/customers/{id}": {
      "get": {
        "description": "Renvoie un client. Réponse 404 si le client n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Client introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir un client par ID",
        "tags": ["Clients"]
      },
      "patch": {
        "description": "Met à jour un client. N'envoyez que les champs à modifier. Réponse 404 si le client n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "whatsapp_number": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "postal_code": {
                    "type": "string"
                  },
                  "is_business": {
                    "type": "boolean"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données d'entrée invalides"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Client introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Mettre à jour un client",
        "tags": ["Clients"]
      },
      "delete": {
        "description": "Retire un client de l'usage actif : il n'apparaît plus dans les listes ni les réponses de détail. Réponse 404 si le client n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client supprimé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Client supprimé avec succès"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Client introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Supprimer un client",
        "tags": ["Clients"]
      }
    },
    "/customers/{id}/transactions": {
      "get": {
        "description": "Renvoie les transactions d'un client. Réponse 404 si le client n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "CustomersController_getTransactions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des transactions du client",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Client introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Transactions du client",
        "tags": ["Clients"]
      }
    },
...
    "/subscriptions/customer/{customerId}": {
      "get": {
        "description": "Renvoie les abonnements d'un client. Réponse 404 si le client n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "SubscriptionsController_findByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "description": "UUID du client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des abonnements du client",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Client introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Abonnements d’un client",
        "tags": ["Abonnements"]
      }
    },
...
  },
  "info": {
    "title": "API lomi.",
    "description": "API de traitement des paiements pour les entreprises d'Afrique de l'Ouest francophone.",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.lomi.africa",
      "description": "Live"
    },
    {
      "url": "https://sandbox.api.lomi.africa",
      "description": "Test"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "schemas": {
...
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Unique transaction identifier"
          },
          "organization_id": {
            "type": "string",
            "example": "789e0123-e89b-12d3-a456-426614174000",
            "description": "Organization ID"
          },
          "customer_id": {
            "type": "string",
            "example": "456e7890-e89b-12d3-a456-426614174000",
            "description": "Customer ID"
          },
          "product_id": {
            "type": "string",
            "example": "321e4567-e89b-12d3-a456-426614174000",
            "description": "Product ID (if applicable)",
            "nullable": true
          },
          "subscription_id": {
            "type": "string",
            "example": "654e7890-e89b-12d3-a456-426614174000",
            "description": "Subscription ID (if applicable)",
            "nullable": true
          },
          "price_id": {
            "type": "string",
            "example": "987e6543-e89b-12d3-a456-426614174000",
            "description": "Price ID used for this transaction",
            "nullable": true
          },
          "transaction_type": {
            "type": "string",
            "example": "payment",
            "description": "Transaction type",
            "enum": ["payment", "refund", "payout"]
          },
          "status": {
            "type": "string",
            "example": "completed",
            "description": "Transaction status",
            "enum": [
              "pending",
              "completed",
              "failed",
              "refunded",
              "expired",
              "cancelled"
            ]
          },
          "description": {
            "type": "string",
            "example": "Payment for Premium Subscription",
            "description": "Transaction description",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Quantity of items"
          },
          "metadata": {
            "type": "object",
            "example": {
              "order_number": "ORD-12345"
            },
            "description": "Additional metadata as JSON",
            "nullable": true
          },
          "gross_amount": {
            "type": "number",
            "example": 10000,
            "description": "Gross amount (total including fees)"
          },
          "discount_amount": {
            "type": "number",
            "example": 500,
            "description": "Total discount amount applied"
          },
          "fee_amount": {
            "type": "number",
            "example": 200,
            "description": "Fee amount charged"
          },
          "net_amount": {
            "type": "number",
            "example": 9300,
            "description": "Net amount (received after fees)"
          },
          "currency_code": {
            "type": "string",
            "example": "XOF",
            "description": "Currency code"
          },
          "provider_code": {
            "type": "string",
            "example": "WAVE",
            "description": "Payment provider code"
          },
          "payment_method_code": {
            "type": "string",
            "example": "MOBILE_MONEY",
            "description": "Payment method code"
          },
          "spi_tx_id": {
            "type": "string",
            "example": "SPI-TX-123456",
            "description": "SPI transaction ID (if applicable)",
            "nullable": true
          },
          "spi_account_number": {
            "type": "string",
            "example": "221771234567",
            "description": "SPI account number used (if applicable)",
            "nullable": true
          },
          "spi_payment_category": {
            "type": "string",
            "example": "000",
            "description": "SPI payment category (if applicable)",
            "nullable": true
          },
          "spi_payment_status": {
            "type": "string",
            "example": "IRREVOCABLE",
            "description": "SPI payment status (if applicable)",
            "nullable": true
          },
          "is_bnpl": {
            "type": "boolean",
            "example": false,
            "description": "Whether this is a Buy Now Pay Later transaction"
          },
          "is_pos": {
            "type": "boolean",
            "example": false,
            "description": "Whether this is a POS (Point of Sale) transaction"
          },
          "spi_date_envoi": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "SPI payment sent date (if applicable)",
            "nullable": true
          },
          "spi_date_irrevocabilite": {
            "type": "string",
            "example": "2024-01-15T10:35:00Z",
            "description": "SPI payment irreversible date (if applicable)",
            "nullable": true
          },
          "environment": {
            "type": "string",
            "example": "live",
            "description": "Environment (test or live)",
            "enum": ["test", "live"]
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the transaction was created"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the transaction was last updated"
          }
        },
        "required": [
          "transaction_id",
          "organization_id",
          "customer_id",
          "product_id",
          "subscription_id",
          "price_id",
          "transaction_type",
          "status",
          "description",
          "quantity",
          "metadata",
          "gross_amount",
          "discount_amount",
          "fee_amount",
          "net_amount",
          "currency_code",
          "provider_code",
          "payment_method_code",
          "spi_tx_id",
          "spi_account_number",
          "spi_payment_category",
          "spi_payment_status",
          "is_bnpl",
          "is_pos",
          "spi_date_envoi",
          "spi_date_irrevocabilite",
          "environment",
          "created_at",
          "updated_at"
        ]
      },
      "CustomerResponseDto": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Unique customer identifier"
          },
          "organization_id": {
            "type": "string",
            "example": "789e0123-e89b-12d3-a456-426614174000",
            "description": "Organization ID the customer belongs to"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "Customer full name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Customer email address",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "example": "+221771234567",
            "description": "Customer phone number",
            "nullable": true
          },
          "whatsapp_number": {
            "type": "string",
            "example": "+221771234567",
            "description": "Customer WhatsApp number",
            "nullable": true
          },
          "country": {
            "type": "string",
            "example": "Senegal",
            "description": "Customer country",
            "nullable": true
          },
          "city": {
            "type": "string",
            "example": "Dakar",
            "description": "Customer city",
            "nullable": true
          },
          "address": {
            "type": "string",
            "example": "123 Main Street",
            "description": "Customer street address",
            "nullable": true
          },
          "postal_code": {
            "type": "string",
            "example": "12345",
            "description": "Customer postal code",
            "nullable": true
          },
          "is_business": {
            "type": "boolean",
            "example": false,
            "description": "Whether the customer is a business"
          },
          "metadata": {
            "type": "object",
            "example": {
              "custom_field": "value"
            },
            "description": "Additional metadata as JSON",
            "nullable": true
          },
          "environment": {
            "type": "string",
            "example": "live",
            "description": "Environment (test or live)",
            "enum": ["test", "live"]
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the customer was created"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the customer was last updated"
          }
        },
        "required": [
          "customer_id",
          "organization_id",
          "name",
          "email",
          "phone_number",
          "whatsapp_number",
          "country",
          "city",
          "address",
          "postal_code",
          "is_business",
          "metadata",
          "environment",
          "created_at",
          "updated_at"
        ]
      },
...
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "subscription_id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Unique subscription identifier"
          },
          "organization_id": {
            "type": "string",
            "example": "789e0123-e89b-12d3-a456-426614174000",
            "description": "Organization ID"
          },
          "product_id": {
            "type": "string",
            "example": "456e7890-e89b-12d3-a456-426614174000",
            "description": "Product ID (recurring product)"
          },
          "price_id": {
            "type": "string",
            "example": "321e4567-e89b-12d3-a456-426614174000",
            "description": "Price ID used for this subscription",
            "nullable": true
          },
          "customer_id": {
            "type": "string",
            "example": "654e7890-e89b-12d3-a456-426614174000",
            "description": "Customer ID"
          },
          "status": {
            "type": "string",
            "example": "active",
            "description": "Subscription status",
            "enum": [
              "pending",
              "active",
              "past_due",
              "cancelled",
              "trialing",
              "paused"
            ]
          },
          "start_date": {
            "type": "string",
            "example": "2024-01-15",
            "description": "Subscription start date"
          },
          "end_date": {
            "type": "string",
            "example": "2024-12-31",
            "description": "Subscription end date (set when cancelled)",
            "nullable": true
          },
          "next_billing_date": {
            "type": "string",
            "example": "2024-02-15",
            "description": "Next billing date (system-managed)",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "example": {
              "notes": "Premium plan"
            },
            "description": "Additional metadata as JSON",
            "nullable": true
          },
          "environment": {
            "type": "string",
            "example": "live",
            "description": "Environment (test or live)",
            "enum": ["test", "live"]
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the subscription was created"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the subscription was last updated"
          }
        },
        "required": [
          "subscription_id",
          "organization_id",
          "product_id",
          "price_id",
          "customer_id",
          "status",
          "start_date",
          "end_date",
          "next_billing_date",
          "metadata",
          "environment",
          "created_at",
          "updated_at"
        ]
      },
...
    }
  }
}